Update matchers for better readability, remove coverage from PRs
